correction

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: dufama <dufama@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/07 14:58:25 by dufama            #+#    #+#             */
-/*   Updated: 2025/10/08 16:28:11 by dufama           ###   ########.fr       */
+/*   Updated: 2025/10/08 19:39:25 by dufama           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,15 +22,15 @@ int	ft_check_args(va_list *args, char c)
 	else if (c == 's')
 		i += ft_putstr_printf(va_arg(*args, char *));
 	else if (c == 'd' || c == 'i')
-		i += ft_putnbr_printf(va_arg(*args, int));
+		i += ft_putnbr_printf((long)va_arg(*args, int));
 	else if (c == 'x')
-		i += ft_putnbr_hex(va_arg(*args, unsigned long));
+		i += ft_putnbr_hex((long)va_arg(*args, unsigned int));
 	else if (c == 'X')
-		i += ft_putnbr_hex_maj(va_arg(*args, unsigned long));
+		i += ft_putnbr_hex_maj((long)va_arg(*args, unsigned int));
 	else if (c == '%')
 		i += ft_putchar_printf('%');
 	else if (c == 'u')
-		i += ft_putnbr_unsigned(va_arg(*args, unsigned long));
+		i += ft_putnbr_unsigned((long)va_arg(*args, unsigned int));
 	else if (c == 'p')
 		i += ft_print_pointer(va_arg(*args, void *));
 	return (i);
